backend/requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1

# AI Integration
openai==1.3.5
anthropic==0.7.7
google-generativeai==0.3.1
requests==2.31.0

# Semantic Search & Analysis (UPDATED for Python 3.13)
sentence-transformers==3.0.0
scikit-learn==1.6.0
numpy==1.26.2
scipy==1.11.4

# Export Features
reportlab==4.0.7
markdown==3.5.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
